feat: спринт 14. попытка 1.

+++ src/main/java/ru/practicum/shareit/exceptions/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.practicum.shareit.exceptions;
 
 public class UserNotFoundException extends NotFoundException {
-    public UserNotFoundException(final int userId) {
+    public UserNotFoundException(final long userId) {
         super("Not found user " + userId);
     }
 }
